refactor(daemon): split local and Herdr spawners
